feat: update frontend, added leaderboard, profile, themes, settings to backend

+++ backend/pkg/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id       int    `json:"-" db:"id"`
+	Id       int64  `json:"-" db:"id"`
 	Name     string `json:"name" binding:"required"`
 	Username string `json:"username" binding:"required"`
 	Password string `json:"password" binding:"required"`
